Bug fixes with Profile Mode.
Fixed a bug when using the default MD5Hash as the name for the profile when saving, and now the profile's name is the same on the folder. (In other words, the appdata folder, instead of being a MD5Hash of the Profile name, will simply be the Profile's name.)

+++ UndertaleModTool/MainWindow.xaml.cs
@@ -3940,7 +3940,7 @@ result in loss of work.");
         {
             if (CurProfileName != "null")
             {
-                var MD5DirName = BitConverter.ToString(MD5CurrentlyLoaded).Replace("-", "").ToLowerInvariant();
+                var MD5DirName = CurProfileName;
                 var MD5DirPath = Path.Combine(ProfilesFolder, MD5DirName);
                 var FileDir = "";
                 string[] iwishiwasbetteratnames = FilePath.Split(new char[] { '\\' });
@@ -3950,11 +3950,11 @@ result in loss of work.");
                     FileDir += iwishiwasbetteratnames[i] + "\\";
                 }
                 FileDir += "Profiles\\" + MD5DirName;
-                Directory.Delete(FileDir, true);
+                if (Directory.Exists(FileDir))
+                    Directory.Delete(FileDir, true);
                 Directory.CreateDirectory(FileDir);
                 DirectoryCopy(MD5DirPath, FileDir, true);
                 this.ShowMessage("Done!");
-                File.WriteAllText(FileDir + "\\" + CurProfileName + ".txt", "This is just to let you know the Profile's name.\nThat's all.");
             }
             else
             {
@@ -3965,7 +3965,7 @@ result in loss of work.");
         {
             if (CanSave == true)
             {
-                var MD5DirName = BitConverter.ToString(MD5CurrentlyLoaded).Replace("-", "").ToLowerInvariant();
+                var MD5DirName = CurProfileName;
                 var FileDir = "";
                 string[] iwishiwasbetteratnames = FilePath.Split(new char[] { '\\' });
                 var directoriesamt = iwishiwasbetteratnames.Length;
